requirements.txt
asttokens==2.4.1
audioread==3.0.1
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==4.4.2
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.9.0
Flask==3.0.0
fonttools==4.46.0
fsspec==2023.4.0
gunicorn==21.2.0
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
ipykernel==6.27.1
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.1
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
more-itertools==10.1.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.7
nest-asyncio==1.5.8
networkx==3.0
noisereduce==3.0.0
numba==0.58.1
numpy==1.26.2
openai-whisper==20231117
packaging==23.2
parso==0.8.3
Pillow==10.1.0
platformdirs==4.1.0
pocketsphinx==5.0.2
pooch==1.8.0
proglog==0.1.10
prompt-toolkit==3.0.43
psutil==5.9.6
pure-eval==0.2.2
PyAudio==0.2.14
pycparser==2.21
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pyzmq==25.1.2
regex==2023.10.3
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sounddevice==0.4.6
soundfile==0.12.1
soxr==0.3.7
stack-data==0.6.3
sympy==1.12
threadpoolctl==3.2.0
tiktoken==0.5.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
typing_extensions==4.4.0
urllib3==2.1.0
wcwidth==0.2.12
Werkzeug==3.0.1